// src/types/product.ts
// A type definition for your product data (adjust as needed)
export interface Product {
    id: string
    name: string
    slug: string
    category: string
    images: string[]
    brand: string
    description: string
    stock: number
    price: string
    rating: string
    numReviews: number
    isFeatured: boolean
    banner: string
    createdAt: string
}